Rename SchnorrSignature to SchnorrSigner
